// MongoAPP/tabs/src/components/models/EventSellerMappingModel.ts
export interface IEventSellerMapping {
    No: number;
    EventName: string;
    CompanyName_CN: string;
    CompanyName_EN: string;
    SellerName_CN: string;
    SellerName_EN: string;
    O365Account: string;
    SellerEmail: string;
    EventOwner: string;
    BookingsID?: string;
    publicUrl?: string;
    contactEmail?: string;
}