// MongoAPP/tabs/src/components/models/SellerBuyerModel.ts
export interface ISellerBuyerModel {
    SellerName: string
    displayName: string;
    emailAddress: string;
    addresses: string;
    phones: string;
}

export interface ISellerModel {
    CompanyName_CN: string;
    CompanyName_EN: string;
    SellerName_CN: string;
    SellerName_EN: string;    
    SellerEmail: string;
}